# the_rivera-master-master/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Citas API REST con Laravel",
        "contact": {
            "email": "s22030116@itsch.edu.mx"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/servicios": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Mostrar todos los servicios",
                "operationId": "3a1c1c392788bd4c8d03626d48d28645",
                "responses": {
                    "200": {
                        "description": "Lista de servicios obtenida correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear un nuevo servicio",
                "operationId": "cd978e7e6c3a84104d6e47bc1775f0f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "precio",
                                    "duracion"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Corte de cabello"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 199.99
                                    },
                                    "duracion": {
                                        "type": "string",
                                        "example": "01:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/servicios/{id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Mostrar un servicio por ID",
                "operationId": "9c04868e405cd028d12617722c41d260",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio encontrado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Actualizar un servicio existente",
                "operationId": "44c627d011d44594d0978735c2198fa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Servicio actualizado"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 300
                                    },
                                    "duracion": {
                                        "type": "string",
                                        "example": "00:45:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Servicio actualizado correctamente"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Eliminar un servicio",
                "operationId": "fdd7b675238376844570194b662ea1da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Servicio eliminado correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer token en el header Authorization"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}